+++ backend/app/main.py
@@ -5,6 +5,8 @@ from starlette.middleware.cors import CORSMiddleware
 from app.db.base import Base, engine
 from app.models import user
 from dotenv import load_dotenv
+from app.api.routers import router
+
 app = FastAPI()
 app.include_router(auth.router)
 app.include_router(posts.router, prefix="/api")
